Refs #16686 Fix merge conflicts
