+++ Framework/CurveFitting/inc/MantidCurveFitting/Functions/DiffSphere.h
@@ -1,16 +1,15 @@
 #ifndef MANTID_DIFFSPHERE_H_
 #define MANTID_DIFFSPHERE_H_
 
-//----------------------------------------------------------------------
-// Includes
-//----------------------------------------------------------------------
-#include "DeltaFunction.h"
-#include "MantidAPI/ParamFunction.h"
-#include "MantidAPI/IFunction1D.h"
-#include "MantidAPI/FunctionDomain.h"
-#include "MantidAPI/Jacobian.h"
+// Mantid Coding standars <http://www.mantidproject.org/Coding_Standards>
+// Mantid Headers from the same project
+#include "MantidCurveFitting/Functions/ElasticDiffSphere.h"
+#include "MantidCurveFitting/Functions/InelasticDiffSphere.h"
+// Mantid headers from other projects
 #include "MantidAPI/ImmutableCompositeFunction.h"
-#include "DeltaFunction.h"
+#include "MantidAPI/IFunction1D.h"
+// third party library headers (N/A)
+// standard library headers (N/A)
 
 namespace Mantid {
 namespace CurveFitting {
@@ -41,97 +40,6 @@ File change history is stored at: <https://github.com/mantidproject/mantid>
 Code Documentation is available at: <http://doxygen.mantidproject.org>
 */
 
-class DLLExport ElasticDiffSphere : public DeltaFunction {
-public:
-  /// Constructor
-  ElasticDiffSphere();
-
-  /// overwrite IFunction base class methods
-  std::string name() const override { return "ElasticDiffSphere"; }
-
-  const std::string category() const override { return "QuasiElastic"; }
-
-  /// A rescaling of the peak intensity
-  double HeightPrefactor() const override;
-
-  /// overwrite IFunction base class method, which declare function parameters
-  void init() override;
-};
-
-/// structure to hold info on Volino's coefficients
-struct xnlc {
-  double x;
-  size_t n;
-  size_t l;
-};
-
-/// simple structure to hold a linear interpolation of factor J around its
-/// numerical divergence point
-struct linearJ {
-  double slope;
-  double intercept;
-};
-
-/* Class representing the inelastic portion of the DiffSphere algorithm.
- * Contains the 98 Lorentzians.
- */
-class DLLExport InelasticDiffSphere : public API::ParamFunction,
-                                      public API::IFunction1D {
-public:
-  InelasticDiffSphere();
-
-  /// overwrite IFunction base class methods
-  void init() override;
-
-  /// overwrite IFunction base class methods
-  std::string name() const override { return "InelasticDiffSphere"; }
-
-  /// overwrite IFunction base class methods
-  const std::string category() const override { return "QuasiElastic"; }
-
-  /// Calculate the (2l+1)*A_{n,l} coefficients for each Lorentzian
-  std::vector<double> LorentzianCoefficients(double a) const;
-
-protected:
-  void function1D(double *out, const double *xValues,
-                  const size_t nData) const override;
-
-private:
-  /// initialize the Xnl coefficients
-  void initXnlCoeff();
-
-  /// initialize the m_alpha coefficients
-  void initAlphaCoeff();
-
-  /// initialize the list of parameters for A_{n,l} linear interpolation around
-  /// the indeterminacy point
-  void initLinJlist();
-
-  /// Cache Q values from the workspace
-  void setWorkspace(boost::shared_ptr<const API::Workspace> ws) override;
-
-  /// xnl coefficients
-  std::vector<xnlc> m_xnl;
-
-  /// certain coefficients invariant during fitting
-  std::vector<double> m_alpha;
-
-  /// maximum value of l in xnlist
-  size_t m_lmax;
-
-  /// linear interpolation zone around the numerical divergence of factor J
-  double m_divZone;
-
-  /// Plank's constant divided by \f$2\pi\f$, in units of meV*THz
-  double m_hbar;
-
-  /// list of linearized J values
-  std::vector<linearJ> m_linearJlist;
-
-  /// list of calculated Q values
-  std::vector<double> m_qValueCache;
-
-}; // end of class InelasticDiffSphere
 
 class DLLExport DiffSphere : public API::ImmutableCompositeFunction {
 
@@ -153,15 +61,16 @@ public:
                                 const API::IFunction::Attribute &defaultValue);
 
   /// Override parent definition
-  void setAttribute(const std::string &name, const Attribute &att) override;
+  void setAttribute(const std::string &name,
+    const API::IFunction::Attribute &att) override;
 
   /// overwrite IFunction base class method, which declare function parameters
   void init() override;
 
 private:
-  boost::shared_ptr<ElasticDiffSphere>
+  boost::shared_ptr<Mantid::CurveFitting::Functions::ElasticDiffSphere>
       m_elastic; // elastic intensity of the DiffSphere structure factor
-  boost::shared_ptr<InelasticDiffSphere>
+  boost::shared_ptr<Mantid::CurveFitting::Functions::InelasticDiffSphere>
       m_inelastic; // inelastic intensity of the DiffSphere structure factor
 };
